requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
deepgram-sdk==2.12.0
pyjwt==2.8.0
pydantic==2.5.2
flask-socketio==5.3.6
python-engineio==4.8.0
python-socketio==5.10.0
sentencepiece==0.1.99
protobuf==4.25.1
livekit==0.22.0
types-protobuf>=3.0.0
Werkzeug==2.3.7 